+++ plugins/woocommerce/src/Internal/Abilities/Domain/ProductAbilityTrait.php
@@ -1,0 +1,135 @@
+<?php
+/**
+ * Product ability trait file.
+ */
+
+declare( strict_types=1 );
+
+namespace Automattic\WooCommerce\Internal\Abilities\Domain;
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * Shared product helpers for WooCommerce domain ability definitions.
+ */
+trait ProductAbilityTrait {
+
+	/**
+	 * Get the shared product mutation input schema.
+	 *
+	 * @return array
+	 */
+	protected static function get_product_mutation_input_schema(): array {
+		return array(
+			'type'                 => 'object',
+			'properties'           => array(
+				'type'              => array( 'type' => 'string' ),
+				'name'              => array( 'type' => 'string' ),
+				'sku'               => array( 'type' => 'string' ),
+				'regular_price'     => array( 'type' => 'string' ),
+				'sale_price'        => array( 'type' => 'string' ),
+				'description'       => array( 'type' => 'string' ),
+				'short_description' => array( 'type' => 'string' ),
+				'status'            => array( 'type' => 'string' ),
+				'manage_stock'      => array( 'type' => 'boolean' ),
+				'stock_quantity'    => array( 'type' => 'integer' ),
+				'stock_status'      => array( 'type' => 'string' ),
+				'virtual'           => array( 'type' => 'boolean' ),
+				'downloadable'      => array( 'type' => 'boolean' ),
+			),
+			'required'             => array( 'name' ),
+			'additionalProperties' => false,
+		);
+	}
+
+	/**
+	 * Get a product from ability input.
+	 *
+	 * @param array $input Ability input.
+	 * @return \WC_Product|\WP_Error
+	 */
+	protected static function get_product_from_input( array $input ) {
+		if ( empty( $input['id'] ) ) {
+			return new \WP_Error(
+				'woocommerce_product_id_required',
+				__( 'Product ID is required.', 'woocommerce' ),
+				array( 'status' => 400 )
+			);
+		}
+
+		$product = wc_get_product( absint( $input['id'] ) );
+
+		if ( ! $product ) {
+			return new \WP_Error(
+				'woocommerce_product_not_found',
+				__( 'Product not found.', 'woocommerce' ),
+				array( 'status' => 404 )
+			);
+		}
+
+		return $product;
+	}
+
+	/**
+	 * Set supported product properties from ability input.
+	 *
+	 * @param \WC_Product $product Product object.
+	 * @param array       $input   Ability input.
+	 */
+	protected static function set_product_props( \WC_Product $product, array $input ): void {
+		$setters = array(
+			'name'              => 'set_name',
+			'sku'               => 'set_sku',
+			'regular_price'     => 'set_regular_price',
+			'sale_price'        => 'set_sale_price',
+			'description'       => 'set_description',
+			'short_description' => 'set_short_description',
+			'status'            => 'set_status',
+			'manage_stock'      => 'set_manage_stock',
+			'stock_quantity'    => 'set_stock_quantity',
+			'stock_status'      => 'set_stock_status',
+			'virtual'           => 'set_virtual',
+			'downloadable'      => 'set_downloadable',
+		);
+
+		foreach ( $setters as $field => $setter ) {
+			if ( array_key_exists( $field, $input ) && is_callable( array( $product, $setter ) ) ) {
+				$value = $input[ $field ];
+
+				if ( is_string( $value ) ) {
+					$value = wc_clean( wp_unslash( $value ) );
+				}
+
+				$product->{$setter}( $value );
+			}
+		}
+	}
+
+	/**
+	 * Format a product for ability output.
+	 *
+	 * @param \WC_Product $product Product object.
+	 * @return array
+	 */
+	protected static function format_product( \WC_Product $product ): array {
+		return array(
+			'id'             => $product->get_id(),
+			'name'           => $product->get_name(),
+			'slug'           => $product->get_slug(),
+			'permalink'      => $product->get_permalink(),
+			'type'           => $product->get_type(),
+			'status'         => $product->get_status(),
+			'sku'            => $product->get_sku(),
+			'price'          => $product->get_price(),
+			'regular_price'  => $product->get_regular_price(),
+			'sale_price'     => $product->get_sale_price(),
+			'stock_status'   => $product->get_stock_status(),
+			'stock_quantity' => null === $product->get_stock_quantity() ? null : (int) $product->get_stock_quantity(),
+			'manage_stock'   => (bool) $product->get_manage_stock(),
+			'virtual'        => (bool) $product->get_virtual(),
+			'downloadable'   => (bool) $product->get_downloadable(),
+			'date_created'   => self::format_datetime( $product->get_date_created() ),
+			'date_modified'  => self::format_datetime( $product->get_date_modified() ),
+		);
+	}
+}
