+++ plugins/woocommerce/src/Internal/Abilities/Domain/ProductUpdate.php
@@ -16,31 +16,39 @@ defined( 'ABSPATH' ) || exit;
  */
 class ProductUpdate extends DomainAbility implements AbilityDefinition {
 
+	use ProductAbilityTrait;
+
 	/**
-	 * Register the ability.
+	 * Get the ability name.
+	 *
+	 * @return string
 	 *
 	 * @since 10.9.0
 	 */
-	public static function register(): void {
-		if ( self::has_ability( 'woocommerce/product-update' ) ) {
-			return;
-		}
+	public static function get_name(): string {
+		return 'woocommerce/product-update';
+	}
 
-		wp_register_ability(
-			'woocommerce/product-update',
-			array(
-				'label'               => __( 'Update Product', 'woocommerce' ),
-				'description'         => __(
-					'Update an existing WooCommerce product using WooCommerce product APIs.',
-					'woocommerce'
-				),
-				'category'            => 'woocommerce',
-				'input_schema'        => self::get_input_schema(),
-				'output_schema'       => self::get_entity_output_schema( 'product' ),
-				'execute_callback'    => array( __CLASS__, 'execute' ),
-				'permission_callback' => array( __CLASS__, 'can_update_product' ),
-				'meta'                => self::get_domain_meta( 'update', false, false, true ),
-			)
+	/**
+	 * Get the ability registration arguments.
+	 *
+	 * @return array
+	 *
+	 * @since 10.9.0
+	 */
+	public static function get_args(): array {
+		return array(
+			'label'               => __( 'Update Product', 'woocommerce' ),
+			'description'         => __(
+				'Update an existing WooCommerce product using WooCommerce product APIs.',
+				'woocommerce'
+			),
+			'category'            => 'woocommerce',
+			'input_schema'        => self::get_input_schema(),
+			'output_schema'       => self::get_entity_output_schema( 'product' ),
+			'execute_callback'    => array( __CLASS__, 'execute' ),
+			'permission_callback' => array( __CLASS__, 'can_update_product' ),
+			'meta'                => self::get_domain_meta( 'update', false, false, true ),
 		);
 	}
